Remove Neoforge support and downgrade Java compatibility level to 17
This commit removes the support for the Neoforge platform from the project and downgrades the Java compatibility level from 21 to 17. Multiple files related to Neoforge have been deleted, and all references to it from code and properties files have been removed. Further, the Java compatibility level for all mixins has been updated to 17.